feat: set id for VIP-298 and add submitted multisig TX
