+++ multisig/simulations/ethereum/index.ts
@@ -21,3 +21,4 @@ import "./vip-021";
 import "./vip-022";
 import "./vip-023";
 import "./vip-024";
+import "./vip-025";
